General Update

+++ Services/Order/Core/MultiShop.Order.Application/Features/CQRS/Commands/OrderDetailCommands/CreateOrderDetailCommand.cs
@@ -12,7 +12,7 @@ namespace MultiShop.Order.Application.Features.CQRS.Commands.OrderDetailCommands
         public string ProductName { get; set; }
         public decimal ProductPrice { get; set; }
         public int ProductAmount { get; set; }
-        public int ProductTotalPrice { get; set; }
+        public decimal ProductTotalPrice { get; set; }
         public int OrderingId { get; set; }
     }
 }
