+++ Services/Order/Core/MultiShop.Order.Application/Features/CQRS/Results/OrderDetailResults/GetOrderDetailQueryResult.cs
@@ -13,7 +13,7 @@ namespace MultiShop.Order.Application.Features.CQRS.Results.OrderDetailResults
         public string ProductName { get; set; }
         public decimal ProductPrice { get; set; }
         public int ProductAmount { get; set; }
-        public int ProductTotalPrice { get; set; }
+        public decimal ProductTotalPrice { get; set; }
         public int OrderingId { get; set; }
     }
 }
